Implement GroovePlaylistModel.
This involved splitting GrooveSearchModel into a base GrooveSongsModel, and
inheriting from that.

GrooveSongsModel's job is to manage the list of songs inside it (and common
actions). GrooveSearchModel now just implements a search API on top of this,
and likewise, GroovePlaylistModel just implements a playlist API on top.

Some rejuggling of tests is probably a good idea; will come in next commit.